web/price_checker/src/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-MySQLdb==2.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
PyJWT==1.7.1
python-crontab==3.2.0
jsonschema==4.23.0
uwsgi
